Bug fixes/enhancements
- Increased font size on display
- Fixed issue where database connections were not closing.  New
connections were being established with every query, and eventually it
was maxing out the connection pool on the server.  Now there will be
only 1 connection per instance of application.
- Added code to handle additional column in ORDERS table.  This change
is in conjunction with change put into Chromis POS.  In order to upgrade
to this version of Chromis Kitchen Screen, Chromis POS needs to be
updated with Nick Deppe's changes for Chromis POS 0.56.3.